+++ spec/dummy/app/jobs/application_job.rb
@@ -1,5 +1,5 @@
 class ApplicationJob < ActiveJob::Base
-  active_job_channel global_broadcast: false
+  active_job_channel public_broadcast: false
 
   def perform(identifier = nil)
     @ajc_identifier = identifier